move metrics components to playground-ui
