+++ packages/playground-ui/package.json
@@ -106,6 +106,7 @@
     "@mastra/client-js": "workspace:^",
     "@mastra/core": ">=1.7.1-0 <2.0.0-0",
     "@mastra/react": "workspace:*",
+    "@tanstack/react-query": "^5.90.21",
     "lucide-react": "^0.474.0",
     "react": ">=19.0.0",
     "tailwindcss": "^4.0.0"
@@ -117,6 +118,7 @@
     "@mastra/react": "workspace:*",
     "@storybook/addon-docs": "^9.1.20",
     "@storybook/react-vite": "^9.1.20",
+    "@tanstack/react-query": "^5.90.21",
     "@tailwindcss/postcss": "^4.2.2",
     "@tailwindcss/typography": "^0.5.19",
     "@tailwindcss/vite": "4.2.2",
