+++ packages/playground-ui/src/domains/metrics/components/date-range-selector.tsx
@@ -1,0 +1,16 @@
+import { SelectFieldBlock } from '../../../ds/components/FormFieldBlocks';
+import { DATE_PRESETS, useMetrics } from '../hooks/use-metrics';
+
+export function DateRangeSelector() {
+  const { datePreset, setDatePreset } = useMetrics();
+
+  return (
+    <SelectFieldBlock
+      name="date-range"
+      labelIsHidden
+      value={datePreset}
+      options={DATE_PRESETS.map(p => ({ label: p.label, value: p.value }))}
+      onValueChange={value => setDatePreset(value as typeof datePreset)}
+    />
+  );
+}
