+++ packages/playground-ui/src/domains/metrics/hooks/use-agent-runs-kpi-metrics.tsx
@@ -1,7 +1,7 @@
 import { useMastraClient } from '@mastra/react';
 import { useQuery } from '@tanstack/react-query';
-
 import { useMetricsFilters } from './use-metrics-filters';
+
 
 /** Total Agent Runs — count of agent duration metric observations */
 export function useAgentRunsKpiMetrics() {
