+++ packages/playground-ui/src/domains/metrics/hooks/use-avg-score-kpi-metrics.tsx
@@ -1,7 +1,7 @@
 import { useMastraClient } from '@mastra/react';
 import { useQuery } from '@tanstack/react-query';
-
 import { useMetricsFilters } from './use-metrics-filters';
+
 
 /** Avg Score — average score across all scorers via aggregate API */
 export function useAvgScoreKpiMetrics() {
