+++ packages/playground-ui/src/domains/metrics/hooks/use-metrics.tsx
@@ -1,4 +1,4 @@
-import { differenceInDays, format } from 'date-fns';
+import { format } from 'date-fns';
 import { createContext, useContext, useState, useCallback, useEffect } from 'react';
 import type { ReactNode } from 'react';
 
@@ -20,40 +20,7 @@ export function isValidPreset(value: string | null | undefined): value is DatePr
   return typeof value === 'string' && (VALID_PRESETS.has(value) || value === 'custom');
 }
 
-const PRESET_DAYS: Record<string, number> = {
-  '24h': 1,
-  '3d': 3,
-  '7d': 7,
-  '14d': 14,
-  '30d': 30,
-};
-
 export type DateRange = { from?: Date; to?: Date };
-export type Comparator = 'is' | 'is not';
-export type FilterGroup = { id: string; field: string; comparator: Comparator; values: string[] };
-
-const ENV_PCTS: Record<string, number> = {
-  'Studio Cloud': 42,
-  Production: 31,
-  Staging: 18,
-  Dev: 7,
-  'CI / Preview': 2,
-};
-
-function getMultiplier(preset: DatePreset, customRange: DateRange | undefined, filterGroups: FilterGroup[]): number {
-  let dateMul = 1;
-  if (preset !== 'custom') {
-    dateMul = PRESET_DAYS[preset] ?? 1;
-  } else if (customRange?.from && customRange?.to) {
-    dateMul = Math.max(1, differenceInDays(customRange.to, customRange.from) + 1);
-  }
-
-  const envGroups = filterGroups.filter(g => g.field === 'Environment' && g.comparator === 'is');
-  const envPct =
-    envGroups.length === 0 ? 100 : envGroups.flatMap(g => g.values).reduce((s, v) => s + (ENV_PCTS[v] ?? 0), 0);
-
-  return dateMul * (envPct / 100);
-}
 
 export const MetricsContext = createContext<{
   datePreset: DatePreset;
@@ -61,18 +28,12 @@ export const MetricsContext = createContext<{
   customRange: DateRange | undefined;
   setCustomRange: (v: DateRange | undefined) => void;
   dateRangeLabel: string;
-  filterGroups: FilterGroup[];
-  setFilterGroups: React.Dispatch<React.SetStateAction<FilterGroup[]>>;
-  multiplier: number;
 }>({
   datePreset: '24h',
   setDatePreset: () => {},
   customRange: undefined,
   setCustomRange: () => {},
   dateRangeLabel: 'Last 24 hours',
-  filterGroups: [],
-  setFilterGroups: () => {},
-  multiplier: 1,
 });
 
 export function useMetrics() {
@@ -85,7 +46,7 @@ function getDateRangeLabel(preset: DatePreset, customRange: DateRange | undefine
   }
   if (customRange?.from) {
     if (customRange.to) {
-      return `${format(customRange.from, 'MMM d, yyyy')} \u2013 ${format(customRange.to, 'MMM d, yyyy')}`;
+      return `${format(customRange.from, 'MMM d, yyyy')} – ${format(customRange.to, 'MMM d, yyyy')}`;
     }
     return format(customRange.from, 'MMM d, yyyy');
   }
@@ -103,11 +64,8 @@ export function MetricsProvider({
 }) {
   const [datePreset, setDatePresetState] = useState<DatePreset>(initialPreset ?? '24h');
   const [customRange, setCustomRange] = useState<DateRange | undefined>(undefined);
-  const [filterGroups, setFilterGroups] = useState<FilterGroup[]>([]);
   const dateRangeLabel = getDateRangeLabel(datePreset, customRange);
-  const multiplier = getMultiplier(datePreset, customRange, filterGroups);
 
-  // Sync from external source (e.g. URL) when initialPreset changes
   useEffect(() => {
     if (initialPreset && initialPreset !== datePreset) {
       setDatePresetState(initialPreset);
@@ -130,9 +88,6 @@ export function MetricsProvider({
         customRange,
         setCustomRange,
         dateRangeLabel,
-        filterGroups,
-        setFilterGroups,
-        multiplier,
       }}
     >
       {children}
