+++ packages/playground-ui/src/domains/metrics/hooks/use-model-cost-kpi-metrics.tsx
@@ -1,7 +1,7 @@
 import { useMastraClient } from '@mastra/react';
 import { useQuery } from '@tanstack/react-query';
-
 import { useMetricsFilters } from './use-metrics-filters';
+
 
 /** Total Model Cost — sum of estimatedCost across input and output token metrics */
 export function useModelCostKpiMetrics() {
