+++ packages/playground-ui/src/domains/metrics/hooks/use-total-tokens-kpi-metrics.tsx
@@ -1,7 +1,7 @@
 import { useMastraClient } from '@mastra/react';
 import { useQuery } from '@tanstack/react-query';
-
 import { useMetricsFilters } from './use-metrics-filters';
+
 
 /** Total Tokens — sum of all input + output tokens */
 export function useTotalTokensKpiMetrics() {
