+++ packages/playground-ui/src/domains/metrics/hooks/use-trace-volume-metrics.tsx
@@ -1,6 +1,5 @@
 import { useMastraClient } from '@mastra/react';
 import { useQuery } from '@tanstack/react-query';
-
 import { useMetricsFilters } from './use-metrics-filters';
 
 export interface VolumeRow {
