+++ packages/playground-ui/src/domains/metrics/index.ts
@@ -1,1 +1,2 @@
 export * from './components';
+export * from './hooks';
