+++ packages/playground-ui/src/index.ts
@@ -145,5 +145,8 @@ export * from './lib/template';
 // Store
 export { usePlaygroundStore, useIsDarkMode, type PlaygroundTheme } from './store/playground-store';
 
+// Domains
+export * from './domains/metrics';
+
 // DS Types
 export type { LinkComponent, LinkComponentProps } from './ds/types/link-component';
