--- packages/playground/src/domains/metrics/components/index.ts
@@ -1,30 +1,0 @@
-export { MetricsDashboard } from './metrics-dashboard';
-export { MetricsProvider, useMetrics, isValidPreset } from '../hooks/use-metrics';
-export type { DatePreset } from '../hooks/use-metrics';
-export { useAgentRunsKpiMetrics } from '../hooks/use-agent-runs-kpi-metrics';
-export { useTotalTokensKpiMetrics } from '../hooks/use-total-tokens-kpi-metrics';
-export { useAvgScoreKpiMetrics } from '../hooks/use-avg-score-kpi-metrics';
-export { useMetricsFilters } from '../hooks/use-metrics-filters';
-export { DateRangeSelector } from './date-range-selector';
-export { MetricsCard } from '@mastra/playground-ui';
-export { MetricsDataTable } from '@mastra/playground-ui';
-export { MetricsKpiCard } from '@mastra/playground-ui';
-export { ModelUsageCostCard } from './model-usage-cost-card';
-export { useModelUsageCostMetrics } from '../hooks/use-model-usage-cost-metrics';
-export type { ModelUsageRow } from '../hooks/use-model-usage-cost-metrics';
-export { HorizontalBars } from '@mastra/playground-ui';
-export { MetricsLineChart, type MetricsLineChartSeries } from '@mastra/playground-ui';
-export { MetricsLineChartTooltip } from '@mastra/playground-ui';
-export { TracesVolumeCard } from './traces-volume-card';
-export { useTraceVolumeMetrics } from '../hooks/use-trace-volume-metrics';
-export type { VolumeRow } from '../hooks/use-trace-volume-metrics';
-export { LatencyCard } from './latency-card';
-export { useLatencyMetrics } from '../hooks/use-latency-metrics';
-export type { LatencyPoint } from '../hooks/use-latency-metrics';
-export { ScoresCard } from './scores-card';
-export { useScoresMetrics } from '../hooks/use-scores-metrics';
-export type { ScorerSummary, ScoresOverTimePoint } from '../hooks/use-scores-metrics';
-export { TokenUsageByAgentCard } from './token-usage-by-agent-card';
-export { useTokenUsageByAgentMetrics } from '../hooks/use-token-usage-by-agent-metrics';
-export type { TokenUsageByAgentRow } from '../hooks/use-token-usage-by-agent-metrics';
-export { formatCompact, CHART_COLORS } from './metrics-utils';
