+++ packages/playground/src/domains/metrics/components/metrics-kpi-cards.tsx
@@ -1,126 +1,65 @@
-import { MetricsKpiCard } from '@mastra/playground-ui';
-import { useAgentRunsKpiMetrics } from '../hooks/use-agent-runs-kpi-metrics';
-import { useAvgScoreKpiMetrics } from '../hooks/use-avg-score-kpi-metrics';
-import { useModelCostKpiMetrics } from '../hooks/use-model-cost-kpi-metrics';
-import { useTotalTokensKpiMetrics } from '../hooks/use-total-tokens-kpi-metrics';
-import { formatCompact, formatCost } from './metrics-utils';
+import {
+  KpiCardView,
+  formatCompact,
+  formatCost,
+  useAgentRunsKpiMetrics,
+  useAvgScoreKpiMetrics,
+  useModelCostKpiMetrics,
+  useTotalTokensKpiMetrics,
+} from '@mastra/playground-ui';
 
 export function AgentRunsKpiCard() {
-  const { data: agentRunsKpi, isLoading, isError } = useAgentRunsKpiMetrics();
-  const hasData = agentRunsKpi?.value != null;
-
+  const { data, isLoading, isError } = useAgentRunsKpiMetrics();
   return (
-    <MetricsKpiCard>
-      <MetricsKpiCard.Label>Total Agent Runs</MetricsKpiCard.Label>
-      <MetricsKpiCard.Value className={hasData ? undefined : 'invisible'}>
-        {hasData ? agentRunsKpi.value!.toLocaleString() : '—'}
-      </MetricsKpiCard.Value>
-      {isError ? (
-        <MetricsKpiCard.Error />
-      ) : isLoading ? (
-        <MetricsKpiCard.Loading />
-      ) : hasData ? (
-        agentRunsKpi.changePercent != null ? (
-          <MetricsKpiCard.Change
-            changePct={agentRunsKpi.changePercent}
-            prevValue={agentRunsKpi.previousValue?.toLocaleString()}
-          />
-        ) : (
-          <MetricsKpiCard.NoChange />
-        )
-      ) : (
-        <MetricsKpiCard.NoData />
-      )}
-    </MetricsKpiCard>
+    <KpiCardView
+      label="Total Agent Runs"
+      value={data?.value != null ? data.value.toLocaleString() : null}
+      prevValue={data?.previousValue != null ? data.previousValue.toLocaleString() : undefined}
+      changePct={data?.changePercent ?? null}
+      isLoading={isLoading}
+      isError={isError}
+    />
   );
 }
 
 export function ModelCostKpiCard() {
-  const { data: costKpi, isLoading, isError } = useModelCostKpiMetrics();
-  const hasData = costKpi?.cost != null;
-
+  const { data, isLoading, isError } = useModelCostKpiMetrics();
   return (
-    <MetricsKpiCard>
-      <MetricsKpiCard.Label>Total Model Cost</MetricsKpiCard.Label>
-      <MetricsKpiCard.Value className={hasData ? undefined : 'invisible'}>
-        {hasData ? formatCost(costKpi.cost!, costKpi.costUnit) : '—'}
-      </MetricsKpiCard.Value>
-      {isError ? (
-        <MetricsKpiCard.Error />
-      ) : isLoading ? (
-        <MetricsKpiCard.Loading />
-      ) : hasData ? (
-        costKpi.costChangePercent != null ? (
-          <MetricsKpiCard.Change
-            changePct={costKpi.costChangePercent}
-            prevValue={costKpi.previousCost != null ? formatCost(costKpi.previousCost, costKpi.costUnit) : undefined}
-          />
-        ) : (
-          <MetricsKpiCard.NoChange />
-        )
-      ) : (
-        <MetricsKpiCard.NoData />
-      )}
-    </MetricsKpiCard>
+    <KpiCardView
+      label="Total Model Cost"
+      value={data?.cost != null ? formatCost(data.cost, data.costUnit) : null}
+      prevValue={data?.previousCost != null ? formatCost(data.previousCost, data.costUnit) : undefined}
+      changePct={data?.costChangePercent ?? null}
+      isLoading={isLoading}
+      isError={isError}
+    />
   );
 }
 
 export function TotalTokensKpiCard() {
-  const { data: totalTokensKpi, isLoading, isError } = useTotalTokensKpiMetrics();
-  const hasData = totalTokensKpi?.value != null;
-
+  const { data, isLoading, isError } = useTotalTokensKpiMetrics();
   return (
-    <MetricsKpiCard>
-      <MetricsKpiCard.Label>Total Tokens</MetricsKpiCard.Label>
-      <MetricsKpiCard.Value className={hasData ? undefined : 'invisible'}>
-        {hasData ? formatCompact(totalTokensKpi.value!) : '—'}
-      </MetricsKpiCard.Value>
-      {isError ? (
-        <MetricsKpiCard.Error />
-      ) : isLoading ? (
-        <MetricsKpiCard.Loading />
-      ) : hasData ? (
-        totalTokensKpi.changePercent != null ? (
-          <MetricsKpiCard.Change
-            changePct={totalTokensKpi.changePercent}
-            prevValue={totalTokensKpi.previousValue != null ? formatCompact(totalTokensKpi.previousValue) : undefined}
-          />
-        ) : (
-          <MetricsKpiCard.NoChange />
-        )
-      ) : (
-        <MetricsKpiCard.NoData />
-      )}
-    </MetricsKpiCard>
+    <KpiCardView
+      label="Total Tokens"
+      value={data?.value != null ? formatCompact(data.value) : null}
+      prevValue={data?.previousValue != null ? formatCompact(data.previousValue) : undefined}
+      changePct={data?.changePercent ?? null}
+      isLoading={isLoading}
+      isError={isError}
+    />
   );
 }
 
 export function AvgScoreKpiCard() {
-  const { data: avgScoreKpi, isLoading, isError } = useAvgScoreKpiMetrics();
-  const hasData = avgScoreKpi?.value != null;
-
+  const { data, isLoading, isError } = useAvgScoreKpiMetrics();
   return (
-    <MetricsKpiCard>
-      <MetricsKpiCard.Label>Avg Score</MetricsKpiCard.Label>
-      <MetricsKpiCard.Value className={hasData ? undefined : 'invisible'}>
-        {hasData ? String(avgScoreKpi.value) : '—'}
-      </MetricsKpiCard.Value>
-      {isError ? (
-        <MetricsKpiCard.Error />
-      ) : isLoading ? (
-        <MetricsKpiCard.Loading />
-      ) : hasData ? (
-        avgScoreKpi.changePercent != null ? (
-          <MetricsKpiCard.Change
-            changePct={avgScoreKpi.changePercent}
-            prevValue={avgScoreKpi.previousValue != null ? String(avgScoreKpi.previousValue) : undefined}
-          />
-        ) : (
-          <MetricsKpiCard.NoChange />
-        )
-      ) : (
-        <MetricsKpiCard.NoData />
-      )}
-    </MetricsKpiCard>
+    <KpiCardView
+      label="Avg Score"
+      value={data?.value != null ? String(data.value) : null}
+      prevValue={data?.previousValue != null ? String(data.previousValue) : undefined}
+      changePct={data?.changePercent ?? null}
+      isLoading={isLoading}
+      isError={isError}
+    />
   );
 }
